# api-specification.yaml
openapi: 3.0.0
info:
  title: CloudBOX API
  version: 1.0.0
  description: 클라우드 스토리지 서비스를 위한 REST API

paths:
  /auth:
    post:
      summary: 사용자 로그인
      tags: [인증]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: 로그인 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string

  /files:
    get:
      summary: 파일 목록 조회
      tags: [파일]
      parameters:
        - name: path
          in: query
          schema:
            type: string
          description: 조회할 경로
      responses:
        200:
          description: 파일 목록 반환 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/File'

    post:
      summary: 파일 업로드
      tags: [파일]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                path:
                  type: string
      responses:
        201:
          description: 파일 업로드 성공

  /files/{fileId}:
    get:
      summary: 파일 다운로드
      tags: [파일]
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 파일 다운로드 성공
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        404:
          description: 파일을 찾을 수 없음
    
    delete:
      summary: 파일 삭제
      tags: [파일]
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: 파일 삭제 성공

  /folders:
    post:
      summary: 폴더 생성
      tags: [폴더]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                path:
                  type: string
      responses:
        201:
          description: 폴더 생성 성공